create a interface user + CORS + new Entities and DTO

+++ Application/Features/Orders/DTOs/CreateOrderRequest.cs
@@ -8,7 +8,7 @@ public class CreateOrderRequest
   public string? CustomerId { get; set; }
   public DateTime? EventDate { get; set; }
   public StatusOrder Status { get; set; } = StatusOrder.Pendente;
-  public int? Sinal { get; set; }
+  public decimal? Sinal { get; set; }
   public decimal? TotalValue { get; set; }
   public List<CreateOrderItemRequest> Items { get; set; } = new();
 }
